adjust some functions between markdown and markup
